refactor to use a more modular file structure
