+++ lib/schwarznasenschaf/core/events/ready.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Schwarznasenschaf::Core::Events::ReadyEvent
+  extend Discordrb::EventContainer
+
+  ready do |event|
+    event.bot.listening = '?help'
+  end
+end
